Implement ExecuteP, ExecuteWithEnv, and ExecutePWithEnv
Execute and ExecuteP are convenience wrappers around Golang's subprocess
toolkit. Minimizing repetition, stuttering, etc in code should help
making this installation process as self documenting as possible.

ExecuteWithEnv and ExecutePWithEnv allow you to avoid repetition when
a set of environment variables is empty or repeated.

These functions are also starting to look like the methods on a struct
that encapsultes externally executed commands; it may be helpful for the
purposes of tracking state for individual steps, allowing for
intelligent rollback/failsafes etc.